re-organized component file structure
