+++ cookbook/src/Components/SubComponents/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Logo from '../designs/Logo/CookBookLogo.svg';
+import Logo from '../../designs/Logo/CookBookLogo.svg';
 
 
 const Footer = () => {
